Upload merged translation script
Final upload of the merged 'TranslationSBMLgenerator.py' script on
behalf of Denis. This script merged the work of both Denis and Daniel.
The script is currently configured to produce a very small
"model_toy.xml" as noted within the README; to generate the full file
(**WARNING: over 1GB in size**) delete the comment from the first line.